Separate Start() from others commands in machine package
